### 2024-04-02 d1bdd94
Updated requirements and removed 'faster' code in exercise 3
