+++ requirements.txt
@@ -9,9 +9,7 @@ jupyter==1.0.0
 gymnasium==0.28.1
 pygame==2.4.0
 scikit-learn==1.1.2
-tensorflow==2.12.0
-scikeras==0.10.0
-torch==2.0.0
-stable-baselines3==2.0.0
+torch
+stable-baselines3
 gym==0.21.0
 pyglet==1.5.27
