### 2024-04-12 a31c2f0
Removed strict gym version requirement
