### 2024-04-15 3d71a59
Updates to the exercises for summer term 2024 (#19)
* ex08: transitioned from tf to torch and clean-up

* ex09: transitioned from tf to torch and clean-up

* ex10: transitioned from tf to torch and clean-up for tasks 1+2

* ex11: transition from tf to torch, updating WIP

* moved some cells in the notebook

* Updated requirements and removed 'faster' code in exercise 3

* first version for ex04 with updated function structure

* updated ex05

* drafts for ex4-7

* Finalised exercise 4

* Small fix in template ex. 4

* Final version ex. 5

* Cleared output

* updated ex06, reworked task 2 and 3
[…]
