### 2024-05-31 a9729ab
Updated tutorial video links for exercises 1-6
