### 2024-09-25 916cd57
Update paper.bib
fix one doi
