+++ paper/paper.bib
@@ -223,7 +223,7 @@
   pages = {101964},
   year = {2020},
   issn = {0933-3657},
-  doi = {https://doi.org/10.1016/j.artmed.2020.101964},
+  doi = {10.1016/j.artmed.2020.101964},
   url = {https://www.sciencedirect.com/science/article/pii/S093336572031229X},
   author = {Antonio Coronato and Muddasar Naeem and Giuseppe {De Pietro} and Giovanni Paragliola},
   keywords = {Artificial intelligence, Reinforcement learning, Healthcare, Personalized medicine},
