+++ lecture_slides/tex/Lecture01.tex
@@ -17,7 +17,7 @@
 		\column[T]{0.14\textwidth}
 		\begin{figure}
 			\centering
-				\includegraphics[height=2.75cm]{fig/lec01/Wallscheid.jpg}
+				\includegraphics[height=2.75cm]{fig/lec01/Abdelwanis.jpg}
 				\caption*{Ali \\Abdelwanis}
 		\end{figure}
 			
