+++ lecture_slides/tex/Lecture07.tex
@@ -1,7 +1,5 @@
-\part{Lecture 07: Planning and Learning with Tabular Methods}
-\title[RL Lecture 07]{Lecture 07: Planning and Learning \\with Tabular Methods}  
-\date{}  
-\frame{\titlepage} 
+\section{Planning and learning with tabular methods}
+
 
 %%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%
 %% Recall: RL Agent Taxonomy%%
@@ -18,12 +16,9 @@
 }
 
 %%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%
-\section{Repetition: model-based and model-free RL} 
+\subsection{Repetition: model-based and model-free RL} 
 %%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%
-\begin{frame}
-\frametitle{Table of contents}
-\tableofcontents
-\end{frame}
+
 
 %%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%
 %% Model-based RL %%
@@ -145,7 +140,6 @@
 \begin{itemize}
 	\item Efficiently uses limited amount of experience (e.g., by replay).  
 	\item Allows integration of available a priori knowledge. 
-	\item Learned models might be re-used for other tasks (e.g., monitoring).
 \end{itemize}}
 \vspace{0.25cm}
 \onslide<2->{Pro model-free / direct RL:
@@ -160,12 +154,9 @@
 }
 
 %%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%
-\section{Dyna: integrated planning, acting and learning} 
+\subsection{Dyna: integrated planning, acting and learning} 
 %%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%
-\begin{frame}
-\frametitle{Table of contents}
-\tableofcontents[currentsection]
-\end{frame}
+
 
 %%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%
 %% The General Dyna Architecture (1)%%
@@ -365,12 +356,9 @@ Compared to default Dyna-$Q$ in \algoref{algo:Dyna_Q}, Dyna-$Q$+ contains the fo
 }
 
 %%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%
-\section{Prioritized sweeping} 
+\subsection{Prioritized sweeping} 
 %%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%
-\begin{frame}
-\frametitle{Table of contents}
-\tableofcontents[currentsection]
-\end{frame}
+
 
 %%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%
 %% Background and Idea %%
[…]
