+++ lecture_slides/tex/Lecture04.tex
@@ -1,4 +1,4 @@
-\section{Monte Carlo Methods}
+\section{Monte Carlo methods}
 
 
 %%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%

[…]
