+++ lecture_slides/tex/Summary_Part_One.tex
@@ -1,5 +1,6 @@
 
 \section*{Summary of part I: reinforcement learning in finite state and action spaces}  
+\addsectiontotoc{Test}
 \renewcommand{\thefigure}{S-I.\arabic{figure}} 
 
 %%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%
