+++ .github/workflows/blank.yml
@@ -1,49 +1,0 @@
-# This is a basic workflow to help you get started with Actions
-
-name: Build and render Latex documents to PDF
-
-# Controls when the workflow will run
-on:
-  # Triggers the workflow on push or pull request events but only for the master branch
-  #  and only when there are changes in tex-files
-  push:
-    branches: [ master ]
-    paths:
-      - '**.tex'
-  pull_request:
-    branches: [ master ]
-    paths:
-      - '**.tex'
-
-  # Allows you to run this workflow manually from the Actions tab
-  workflow_dispatch:
-
-# A workflow run is made up of one or more jobs that can run sequentially or in parallel
-jobs:
-  # This workflow contains a single job called "build"
-  build_latex:
-    # The type of runner that the job will run on
-    runs-on: ubuntu-latest
-
-    # Steps represent a sequence of tasks that will be executed as part of the job
-    steps:
-      # Checks-out your repository under $GITHUB_WORKSPACE, so your job can access it
-      - name: Set up Git repository
-        uses: actions/checkout@v2
-
-      - name: Compile LaTeX document
-        uses: xu-cheng/texlive-action/full@v1
-        with:
-          run: |
-            cd lecture_slides
-            python3 build.py
-      - name: Deploy PDFs to server
-        uses: appleboy/scp-action@master
-        with:
-          host: ${{ secrets.UPB_HOST }}
-          username: ${{ secrets.UPB_USER }}
-          password: ${{ secrets.UPB_PW }}
-          port: 22
-          source: "lecture_slides/built/"
-          target: ${{ secrets.LEA_GRP_PATH_RL }}
-
