+++ paper/paper.md
@@ -1,138 +1,0 @@
----
-title: 'Reinforcement Learning: A Comprehensive Open-Source Course'
-tags:
-  - data science
-  - Python
-  - TensorFlow
-  - PyTorch
-  - Jupyter notebook
-  - reproducible workflow
-  - open science
-  - reinforcement learning
-  - exploratory data analysis
-  - machine learning
-  - supervised learning
-authors:
-  - name: Barnabas Haucke-Korber
-    orcid: 0000-0003-0862-2069
-    affiliation: 1
-    
-  - name: Darius Jakobeit
-    orcid: 0009-0002-1576-2465
-    affiliation: 1
-    
-  - name: Wilhelm Kirchgässner
-    orcid: 0000-0001-9490-1843
-    affiliation: 1
-
-  - name: Marvin Meyer
-    orcid: 0009-0008-2879-7118
-    affiliation: 1
-    
-  - name: Maximilian Schenke
-    orcid: 0000-0001-5427-9527
-    affiliation: 1
-        
-  - name: Oliver Wallscheid
-    orcid: 0000-0001-9362-8777
-    affiliation: 1
-
-  - name: Daniel Weber
-    orcid: 0000-0003-3367-5998
-    affiliation: 1
-
-affiliations:
- - name: Department of Power Electronics and Electrical Drives, Paderborn University, Germany
-   index: 1
-
-date: 19 July 2023
-bibliography: paper.bib
----
-
-# Summary
-
-We present an open-source repository of an extensive course on reinforcement learning. It is specifically designed for master students in engineering and computer science. The course aims to introduce beginners to the fundamentals of reinforcement learning and progress towards advanced algorithms. This is done using examples spanning many different classic control engineering tasks. It is structured to be accessible to students with limited prior programming experience by introducing the basics of Python.
-
-The course spans 14 weeks, comprising 14 lectures and 12 exercises. Accompanying video materials from real lectures and exercises are provided to aid in understanding the course content. They are available on the departments' [YouTube channel](https://www.youtube.com/@uni_pb_lea/playlists) under an Creative Commons license. The open-source nature of the course allows other teachers to freely adapt the materials for their own teaching purposes. The primary goal is to equip learners with a solid theory of reinforcement learning principles, as well as the practical tools to solve real-world engineering problems from different domains, such as electrical engineering.
-
-The lecture follows Richard S. Sutton and Andrew G. Barto's fundamentals book on reinforcement learning [@Sutton2005ReinforcementLA] and takes inspiration from the reinforcement learning lecture script delivered by David Silver [@silver2015]. The exercises are programmed in Python using Jupyter notebooks [@Kluyver2016jupyter] for presentation. Important libraries for machine and reinforcement learning are introduced, such as pandas [@reback2020pandas][@mckinney-proc-scipy-2010], gymnasium [@towers_gymnasium_2023], TensorFlow [@tensorflow2015-whitepaper], PyTorch [@NEURIPS2019_9015], scikit-learn [@sklearn_api], and stable-baselines3 [@stable-baselines3]. 
-
-The authors of this course have experience working with reinforcement learning in the domain of electrical engineering, in particular in electric drive [@deep_q_torque] and grid control [@10182718]. The course has first been held under the constraints of the Corona pandemic in 2020, resorting to an online, asynchronous learning experience. It has been extended with a session on more contemporary algorithms in 2022. In 2023, it has been revised and shortened to incorporate experience from teaching the online course and to return to a synchronous and full-presence course format. Both versions (online teaching and offline teaching) are available inside the publicly available [GitHub repository](https://github.com/upb-lea/reinforcement_learning_course_materials).
-
-# Statement of Need
-
[…]
