+++ lecture_slides/tex/Summary_Part_Two.tex
@@ -1,9 +1,8 @@
-%\setcounter{figure}{0} % Manually reset figure count (summary is not a beamer class part)
-%\setcounter{framenumber}{0} % Manually reset page count (summary is not a beamer class part)
-%\pdfbookmark[1]{Summary of Part II: Reinforcement Learning Using Function Approximation}{Sum_Part_Two} 
-\section*{Summary of part II: reinforcement learning using function approximation}  
-%\date{}  
-%\frame{\titlepage}
+{
+	\AtBeginSection[]{} % Deactivate automatic section start slide for this frame
+	\section[Summary of part II]{Summary of part II: continuous state and action spaces}  
+}
+
 \renewcommand{\thefigure}{S-II.\arabic{figure}}  
 
 %%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%
@@ -17,7 +16,7 @@
 	   	\end{varblock}
 	\column{0.5\textwidth}
 		\small
-		  \tableofcontents[hideallsubsections, sections={8-15}]
+		  \tableofcontents[hideallsubsections, sections={9-15}]
 	\end{columns}
 \end{frame}
 
