+++ lecture_slides/main.tex
@@ -4,13 +4,6 @@
 \author{Oliver Wallscheid}
 \date{}
 
-\makeatletter
-\newcommand\addsectiontotoc[1]{%
-\addtocontents{toc}{%
-  \protect\beamer@sectionintoc{\the\c@section}{#1}{\the\c@page}{\the\c@part}%
-    {}}
-}
-\makeatother
 
 %%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%
 %%%% Ohter useful short commands  %%%%%
@@ -34,13 +27,14 @@
 %% Outline / table of content %%
 %%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%
 \begin{frame}{Table of contents}
+  \vspace{-0.2cm}
   \begin{columns}
     \column{0.45\textwidth}
         \small
-        \tableofcontents[hideallsubsections, sections={1-7}]
+        \tableofcontents[hideallsubsections, sections={1-8}]
     \column{0.45\textwidth}
       \small
-        \tableofcontents[hideallsubsections, sections={8-15}]
+        \tableofcontents[hideallsubsections, sections={9-16}]
   \end{columns}
 \end{frame}
 

[…]
