+++ lecture/tex/Lecture14.tex
@@ -87,7 +87,6 @@ Real-world systems are always subject to certain state constraints $\mathcal{X}$
 \centering
 \includegraphics[height=0.68\textheight]{fig/lec14/Safeguard_Classes.pdf}
 \caption{Safety and model knowledge map (derived from L. Brunke et al., \textit{Safe Learning in Robotics: From Learning-Based Control to Safe Reinforcement Learning}, Annual Review of Control, Robotics, and Autonomous Systems, 2022)}
-\label{fig:safety_levels}
 \end{figure}
 }
 
@@ -149,7 +148,6 @@ Real-world systems are always subject to certain state constraints $\mathcal{X}$
 	%\includegraphics[width=1.11\textwidth, height=0.8\textheight]{fig/lec14/Reward_v_mre.pdf}
 	\includegraphics[trim=50 40 10 60, clip, width=1.1\linewidth]{fig/lec14/Reward_v_mre.pdf}
 	\caption{Reward function \ref{eq:reward} for different reference and measured voltages and currents below nominal current}
-	\label{fig:RL_OMG}
 \end{figure}
 \end{minipage}
 \end{column}
@@ -193,7 +191,6 @@ Real-world systems are always subject to certain state constraints $\mathcal{X}$
 \begin{figure}
 	\includegraphics[width=0.9\textwidth, height=0.65\textheight]{fig/lec14/Safe_RL_Shield_OMG.pdf}
 	\caption{Safety shield based on feasible set}
-	\label{fig:RL_OMG}
 \end{figure}
 \end{minipage}
 \end{column}
@@ -222,7 +219,6 @@ Real-world systems are always subject to certain state constraints $\mathcal{X}$
 \begin{figure}
 	\includegraphics[width=0.85\textwidth, height=0.55\textheight]{fig/lec14/OMG_all_aborts_log.pdf}
 	\caption{Accumulated unsafe events (overcurrent/-voltage) per trainingstep $k$}
-	\label{fig:RL_OMG}
 \end{figure}
 \end{minipage}
 \end{column}
@@ -253,7 +249,6 @@ Power Electronic Systems}, 2023
 \begin{figure}
 	\includegraphics[width=0.75\textwidth, height=0.55\textheight]{fig/lec14/370_vDC_SL_RLS_varLoadtimeseries_testcase.pdf}
 	\caption{Blackstart after training using \textcolor{signalgreen}{$\mathrm{DDPG}_\mathrm{SG,RLS}$}}
-	\label{fig:RL_OMG}
 \end{figure}
 \end{minipage}
 \end{column}
