### 2025-03-23 1d2229f
adding automated section labels for pdf destinations
