+++ lecture/main.tex
@@ -12,7 +12,7 @@
 
 %%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%
 %%%Lecture Include Onlys%%%
-%\includeonly{tex/Lecture01, tex/Lecture02, tex/Lecture03, tex/Lecture04, tex/Lecture05, tex/Lecture06, tex/Lecture07, tex/Summary_Part_One, tex/Lecture08}
+\includeonly{tex/Lecture01, tex/Lecture02}
 %%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%
 
 \begin{document}
