+++ README.md
@@ -20,55 +20,55 @@ Lecture notes, tutorial tasks including solutions as well as online videos for a
 ## Lecture slides (click on preview picture)
 <a href="https://ias-uni-siegen.github.io/RL_course/lecture.pdf" target="_blank" class="image fit"><img src="misc/Lecture_preview.jpg" alt=""></a>
 
-01. Introduction to Reinforcement Learning
+01. Introduction to reinforcement learning
     * [Lecture video, part 1](https://www.youtube.com/watch?v=YqlNOCD0rfA)
     * [Lecture video, part 2](https://youtu.be/Yd99sn-64Z8)
-    * [Lecture slides](https://groups.uni-paderborn.de/lea/share/lehre/reinforcementlearning/lecture_slides/built/Lecture01.pdf#page=3)   
-02. Markov Decision Processes
+    * [Lecture slides](https://ias-uni-siegen.github.io/RL_course/lecture.pdf#sec1)   
+02. Markov decision processes
     * [Lecture video](https://www.youtube.com/watch?v=ywn81iGQISE)
-    * [Lecture slides](https://groups.uni-paderborn.de/lea/share/lehre/reinforcementlearning/lecture_slides/built/Lecture02.pdf) 
-03. Dynamic Programming
+    * [Lecture slides](https://ias-uni-siegen.github.io/RL_course/lecture.pdf#sec2) 
+03. Dynamic programming
     * [Lecture video](https://www.youtube.com/watch?v=vjIiYdidFPY)
-    * [Lecture slides](https://groups.uni-paderborn.de/lea/share/lehre/reinforcementlearning/lecture_slides/built/Lecture03.pdf) 
-04. Monte Carlo Methods
+    * [Lecture slides](https://ias-uni-siegen.github.io/RL_course/lecture.pdf#sec3) 
+04. Monte Carlo methods
     * [Lecture video](https://www.youtube.com/watch?v=GBL0ArlONrM)
-    * [Lecture slides](https://groups.uni-paderborn.de/lea/share/lehre/reinforcementlearning/lecture_slides/built/Lecture04.pdf) 
-05. Temporal-Difference Learning
+    * [Lecture slides](https://ias-uni-siegen.github.io/RL_course/lecture.pdf#sec4) 
+05. Temporal-difference learning
     * [Lecture video](https://www.youtube.com/watch?v=Rnf9Wanxnj8)
-    * [Lecture slides](https://groups.uni-paderborn.de/lea/share/lehre/reinforcementlearning/lecture_slides/built/Lecture05.pdf) 
-06. Multi-Step Bootstrapping
+    * [Lecture slides](https://ias-uni-siegen.github.io/RL_course/lecture.pdf#sec5) 
+06. Multi-step bootstrapping
     * [Lecture video](https://www.youtube.com/watch?v=YYTSZTyjbQ4)
-    * [Lecture slides](https://groups.uni-paderborn.de/lea/share/lehre/reinforcementlearning/lecture_slides/built/Lecture06.pdf) 
-07. Planning and Learning with Tabular Methods
+    * [Lecture slides](https://ias-uni-siegen.github.io/RL_course/lecture.pdf#sec6) 
+07. Planning and learning with tabular methods
     * [Lecture video](https://www.youtube.com/watch?v=gvJ3__GmHqo)
-    * [Lecture slides](https://groups.uni-paderborn.de/lea/share/lehre/reinforcementlearning/lecture_slides/built/Lecture07.pdf) 
-08. Function Approximation with Supervised Learning
+    * [Lecture slides](https://ias-uni-siegen.github.io/RL_course/lecture.pdf#sec7) 
+08. Function approximation with supervised learning
     * [Lecture video](https://www.youtube.com/watch?v=tXAxTiuvges)
[…]
