### 2025-04-15 fc9dd60
minor typo corr and animation updates
