+++ lecture/tex/Lecture03.tex
@@ -662,8 +662,8 @@ Assume $\alpha=0.2$ and $\gamma=0.8$ and start with \hl{'tree lover' initial pol
 \frame{\frametitle{Generalized policy iteration (GPI)}
 \begin{itemize}
 	\item Almost all RL methods are well-described as GPI.
-	\item \hl{Push-pull}: Improving the policy will deteriorate value estimation.  
-	\item Well balanced \hl{trade-off between evaluating and improving} is required.
+	\item<2-> \hl{Push-pull}: Improving the policy will deteriorate value estimation.  
+	\item<3-> Well balanced \hl{trade-off between evaluating and improving} is required.
 \end{itemize}
 \begin{figure}
 	\subfloat{
