+++ lecture/tex/Lecture02.tex
@@ -110,18 +110,19 @@ with $p_{ij}\in\left\{\mathbb{R}|0 \leq p_{ij} \leq 1 \right\}$ being the specif
 \end{minipage}
 \begin{minipage}[t]{0.45\linewidth}
 \vspace{-0.7cm}
+\onslide<5->{
 \begin{align*}
    x&\in\left\{1,2,3,4\right\}\\
 				 &=\left\{\mbox{small},\mbox{medium},\mbox{large},\mbox{gone}\right\}\\[1em]
 	\bm{\mathcal{P}} &= \begin{bmatrix}0 & 1-\alpha & 0 & \alpha  \\ 0 & 0 &1-\alpha & \alpha \\ 0 & 0 &1-\alpha & \alpha \\ 0 & 0 & 0 & 1\end{bmatrix}
-\end{align*}
+\end{align*}}
 \end{minipage}
 \begin{itemize}
 	\item At $x=1$ a small tree is planted ('starting point').
-	\item A tree grows with $(1-\alpha)$ probability.
-	\item If it reaches $x=3$ (large) its growth is limited.
-	\item With $\alpha$ probability a natural hazard destroys the tree.
-	\item The state $x=4$ is terminal ('infinite loop').
+	\item<2-> A tree grows with $(1-\alpha)$ probability.
+	\item<3-> With $\alpha$ probability a natural hazard destroys the tree.
+	\item<4-> If it reaches $x=3$ (large) its growth is limited.
+	\item<4-> The state $x=4$ is terminal ('infinite loop').
 \end{itemize}
 }
 
@@ -134,10 +135,10 @@ with $p_{ij}\in\left\{\mathbb{R}|0 \leq p_{ij} \leq 1 \right\}$ being the specif
 \end{figure}
 Possible \hl{samples} for the given Markov chain example starting from $x=1$ (small tree):
 \begin{itemize}
-	\item Small $\rightarrow$ gone
-	\item Small $\rightarrow$ medium $\rightarrow$ gone
-	\item Small $\rightarrow$ medium $\rightarrow$ large $\rightarrow$ gone
-	\item Small $\rightarrow$ medium $\rightarrow$ large $\rightarrow$ large $\rightarrow\ldots$ 
+	\item<1-> Small $\rightarrow$ gone
+	\item<2-> Small $\rightarrow$ medium $\rightarrow$ gone
+	\item<3-> Small $\rightarrow$ medium $\rightarrow$ large $\rightarrow$ gone
+	\item<4-> Small $\rightarrow$ medium $\rightarrow$ large $\rightarrow$ large $\rightarrow\ldots$ 
 \end{itemize}
 }
 
@@ -183,7 +184,7 @@ A \hl{finite Markov reward process (MRP)} is a tuple $\left\langle\mathcal{X}, \
 		\item appreciated by hikers and
 		\item useful for wood production.
 	\end{itemize}
-	\item Loosing a tree due to a hazard is unrewarded.
+	\item<2-> Loosing a tree due to a hazard is unrewarded.
 \end{itemize}
 }
 
@@ -219,14 +220,15 @@ The \hl{state-value function} $v(x_k)$ of an MRP is the expected return starting
 \end{equation}
 \end{defi}
 \begin{itemize}
-	\item Represents the long-term value of being in state $X_k$.
+	\item<2-> Represents the long-term value of being in state $X_k$.
 \end{itemize}
 \vspace{0.25cm}
+\onslide<3->{
 \begin{figure}
 	\includegraphics[height=0.325\textheight]{fig/lec02/Value_Golf.pdf}
[…]
