### 2025-04-18 7f668e3
typo corr and further animations to lec04
