+++ lecture/tex/Lecture04.tex
@@ -16,7 +16,7 @@ Monte Carlo methods:
 \begin{itemize}
 	\item \hl{Model-free} prediction and control
 	\item Estimating value functions and optimize policies in \hl{unknown MDPs} \pause
-	\item But: still assuming finite MDP problems (or problems close to that) \pause
+	\item But: still assuming finite MDP problem structure \pause
 	\item In general: broad class of computational algorithms relying on \hl{repeated random sampling} to obtain numerical results
 \end{itemize}
 }
@@ -27,9 +27,9 @@ Monte Carlo methods:
 \frame{\frametitle{General Monte Carlo (MC) methods' characteristics}
 \begin{itemize}
 	\onslide<1->\item \hl{Learning  from experience}, i.e., sequences of samples $\left\langle x_k, u_k, r_{k+1}\right\rangle$
-	\onslide<2->\item Main concept: Estimation by \hl{averaging sample returns}
+	\onslide<2->\item Main concept: estimation by \hl{averaging sample returns}
 	\onslide<3->\item To guarantee well-defined returns: \hl{limited to episodic tasks}
-	\onslide<4->\item Consequence: Estimation and policy updates only possible in an episode-by-episode way compared to step-by-step (online)
+	\onslide<4->\item Consequence: estimation and policy updates only possible in an episode-by-episode way compared to step-by-step (online)
 \end{itemize}
 \onslide<1->\begin{figure}
 \includegraphics[width=8cm]{fig/lec04/Monte_Carlo_City.jpg}
@@ -103,47 +103,47 @@ $l(x)\leftarrow$ an empty list for every $x\in\mathcal{X}$}
 \frame{\frametitle{Incremental implementation}
 \begin{itemize}
 	\item \algoref{algo:MC_state_prediction_first_visit} is inefficient due to large memory demand.
-	\item Better: use \hl{incremental / recursive implementation}.\pause
-	\item The sample mean $\mu_1,\mu_2,\ldots$ of an arbitrary sequence $g_1, g_2, \ldots$ is:
+	\item Better: use \hl{incremental / recursive implementation}.
+	\item<2-> The sample mean $\mu_1,\mu_2,\ldots$ of an arbitrary sequence $g_1, g_2, \ldots$ is:
 \end{itemize}
 \begin{equation}
 	\begin{split}
-		\mu_J &= \frac{1}{J}\sum_{i=1}^J g_i= \frac{1}{J}\left[g_J + \sum_{i=1}^{J-1} g_i\right]\\
-					&= \frac{1}{J}\left[g_J + (J-1) \mu_{J-1}\right] =\mu_{J-1} + \frac{1}{J}\left[g_J-\mu_{J-1}\right].
+		\onslide<3->{\mu_J &= \frac{1}{J}\sum_{i=1}^J g_i }\onslide<4->{= \frac{1}{J}\left[g_J + \sum_{i=1}^{J-1} g_i\right]}\\
+					&\onslide<5->{= \frac{1}{J}\left[g_J + (J-1) \mu_{J-1}\right]} \onslide<6->{=\mu_{J-1} + \frac{1}{J}\left[g_J-\mu_{J-1}\right].}
 	\end{split}
 	\label{eq:inc_impl_MC_pred}
-\end{equation} \pause
+\end{equation} 
+\onslide<7->{
 \begin{itemize}
-	\item \hl{If a given decision problem is non-stationary}, using a forgetting factor $\alpha\in\left\{\mathbb{R}|0<\alpha<1\right\}$ allows for dynamic adaption: 
[…]
