### 2025-04-18 abb1339
typo corr and animations in lec05
