+++ lecture/tex/Lecture05.tex
@@ -8,8 +8,8 @@
 \frame{\frametitle{Temporal-difference learning and the previous methods}
 \hl{Temporal-difference (TD) learning} combines the previous ideas introduced in DP and MC:
 \begin{itemize}
-	\item From Monte Carlo (MC) methods: Learns directly from experience.
-	\item From dynamic programming (DP): Updates estimates based on other learned estimates (bootstrap).
+	\item From Monte Carlo (MC) methods: learns directly from experience.
+	\item From dynamic programming (DP): updates estimates based on other learned estimates (bootstrap).
 \end{itemize}
 \pause
 \vspace{1cm}
@@ -80,7 +80,7 @@ Recap the every-visit MC update rule \eqref{eq:inc_impl_MC_pred_non_stat} for no
 \label{algo:TD_zero_state_prediction_tabular}
 \end{algorithm}
 \begin{itemize}
-	\item Note that the algorithm can be directly adapted to action-value prediction as it will be used for the later TD-based control approaches.
+	\item<2-> Note that the algorithm can be directly adapted to action-value prediction as it will be used for the later TD-based control approaches.
 \end{itemize}
 }
 
@@ -133,11 +133,10 @@ The \hl{TD error} is:
 																									\onslide<5->{&=\sum_{i=k}^{T-1} \gamma^{i-k}\delta_i.}
 \end{split}
 \end{equation}
-\onslide<5->{
 \begin{itemize}
-	\item MC error is the discounted sum of TD errors in this simplified case.
-	\item If $\hat{v}(x)$ is updated during an episode (as expected in TD(0)), the above identity only holds approximately. 
-\end{itemize}}
+	\item<6-> MC error is the discounted sum of TD errors in this simplified case.
+	\item<7-> If $\hat{v}(x)$ is updated during an episode (as expected in TD(0)), the above identity only holds approximately. 
+\end{itemize}
 }
 
 %%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%
@@ -162,15 +161,15 @@ The \hl{TD error} is:
 \end{figure}
 \vspace{-0.25cm}
 \begin{itemize}
-	\item TD can learn before knowing the final outcome.
+	\item<2-> TD can learn before knowing the final outcome.
 	\begin{itemize}
-		\item TD learns after every step.
-		\item MC must wait until the episode's end.
+		\item<2-> TD learns after every step.
+		\item<3-> MC must wait until the episode's end.
 	\end{itemize}
-	\item TD could learn without a final outcome.
+	\item<4-> TD could learn without a final outcome.
 	\begin{itemize}
-		\item MC is only applicable to episodic tasks.
-		\item TD can learn from incomplete sequences, i.e., in continuing tasks.
+		\item<4-> TD can learn from incomplete sequences, i.e., in continuing tasks.
+		\item<5-> MC is only applicable to episodic tasks.
 	\end{itemize}
 \end{itemize}
 }
@@ -363,7 +362,7 @@ Now consider state A assuming the steady state of batch learning process:
 Looking at \tabref{tab:state_reward_AB_example} there is only one episode visiting state A, where the sample return is $g_{k,j}=0$. Hence, it follows: 
 \begin{equation*}
 	\hat{v}(\mbox{A})|_{\mbox{MC}} =0, \quad \hat{v}(\mbox{A})|_{\mbox{TD}} =\gamma\hat{v}(B)=\frac{3}{4}.
-\end{equation*}
+\end{equation*}\pause
 Where does this mismatch between the MC and TD estimates come from?
 }
 
@@ -382,8 +381,8 @@ Where does this mismatch between the MC and TD estimates come from?
[…]
