### 2025-04-22 6f6efc0
Ex12 write new render method for Goddard environment
