### 2025-04-24 555836c
Ex10 add missing files and remove deprecationWarnings
