### 2025-04-24 b07442d
Ex12 modify rocket environment
