### 2025-04-27 0c222f2
Update README.md
Update to Python 3.12
