### 2025-04-29 fb185b1
minor update to lec06
