+++ lecture/tex/Lecture07.tex
@@ -10,8 +10,8 @@
 	\caption{Main categories of reinforcement learning algorithms\\  \SilverLectureSource}
 \end{figure}
 \begin{itemize}
-	\item Up to now: independent usage of model-free (MC, TD) and model-based RL (DP)
-	\item Today: integrating both strategies (on finite  state \& action spaces)
+	\item \hl{Up to now}: independent usage of model-free (MC, TD) and model-based RL (DP)
+	\item \hl{Today}: integrating both strategies (on finite  state \& action spaces)
 \end{itemize}
 }
 
@@ -116,11 +116,11 @@
 %%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%
 \frame{\frametitle{Model-free RL}
 \begin{itemize}
-	\item \hl{Learn} value functions and/or policy directly from experience.
+	\item \hl{Learn} value functions and/or policy directly from experience (i.e., data).
 	\item Requires no model at all (policy can be considered an implicit model). \pause
 	\item Solves control problems by learning algorithms such as
 	\begin{itemize}
-		\item Monte-Carlo,
+		\item Monte Carlo,
 		\item SARSA or
 		\item $Q$-learning.
 	\end{itemize}
@@ -215,8 +215,8 @@
 				$\hat{q}(x_k, u_k) \leftarrow \hat{q}(x_k, u_k) + \alpha\left[r_{k+1}+\gamma \max_{u} \hat{q}(x_{k+1}, u) - \hat{q}(x_k, u_k)\right]$\;
 				$\hat{\mathcal{M}}(x_k,u_k)\leftarrow \left\{r_{k+1}, x_{k+1}\right\}$ (assuming deterministic env.)\;
 				\For{$i=1,2,\ldots$ n}{
-						$\tilde{x}_i\leftarrow$ random previously visited state\;
-						$\tilde{u}_i\leftarrow$ random previously taken action in $\tilde{x}_i$\;
+						$\tilde{x}_i\leftarrow$ random previously visited or picked state\;
+						$\tilde{u}_i\leftarrow$ random previously taken or picked action in $\tilde{x}_i$\;
 						$\left\{\tilde{r}_{i+1}, \tilde{x}_{i+1}\right\}\leftarrow\hat{\mathcal{M}}(\tilde{x}_i, \tilde{u}_i)$\;
 						$\hat{q}(\tilde{x}_i, \tilde{u}_i) \leftarrow \hat{q}(\tilde{x}_i, \tilde{u}_i) + \alpha\left[\tilde{r}_{i+1}+\gamma \max_{u} \hat{q}(\tilde{x}_{i+1}, u) - \hat{q}(\tilde{x}_i, \tilde{u}_i)\right]$\;
 				}
@@ -236,7 +236,7 @@ The specific Dyna-$Q$ characteristics are:
 \begin{itemize}
 	\item Direct RL update: $Q$-learning, 
 	\item Model: simple memory buffer of previous real experience,
-	\item Search strategy: random choices from model buffer.
+	\item Search strategy: random choices or purposeful selections from model buffer.
 \end{itemize}\pause
 \vspace{0.5cm}
 Moreover:
@@ -292,10 +292,10 @@ Moreover:
 \frame{\frametitle{Maze example (2)}
 \begin{itemize}
 	\item Blocks without an arrow depict a neutral policy (equal action values).
-	\item Black squares indicate agent's position during second episode.
-	\item Without planning ($n=0$), each episodes only adds one new item to the policy.
-	\item With planning ($n=50$), the available experience is efficiently utilized.
-		\item After the third episode, the planning agent found the optimal policy.
[…]
