### 2025-05-26 13b8713
typo corr to lec08
