+++ lecture/tex/Lecture08.tex
@@ -21,66 +21,11 @@ Machine learning (ML) and especially the field of supervised learning (SL) is ex
 	\end{itemize}\pause
 	\item Books classics
 	\begin{itemize}
-		\item \href{http://users.isr.ist.utl.pt/~wurmd/Livros/school/Bishop\%20-\%20Pattern\%20Recognition\%20And\%20Machine\%20Learning\%20-\%20Springer\%20\%202006.pdf}{\textit{Pattern Recognition and Machine Learning} by C. M. Bishop}
-		\item \href{http://www-stat.stanford.edu/~tibs/ElemStatLearn/printings/ESLII_print10.pdf}{\textit{The Elements of Statistical Learning} by Hastie et al.}
-		\item \href{https://www.deeplearningbook.org/}{\textit{Deep Learning} by I. Goodfellow, Y. Bengio, and A. Courville}
+		\item \textit{Pattern Recognition and Machine Learning} by C. M. Bishop
+		\item \textit{The Elements of Statistical Learning} by Hastie et al.
+		\item \textit{Deep Learning} by I. Goodfellow, Y. Bengio, and A. Courville
 	\end{itemize}
 \end{itemize}
-}
-
-
-%  ML hype in industry
-\frame{\frametitle{Machine learning in industry}
-	Machine learning applications are a \href{https://magazine.startus.cc/how-machine-learning-is-changing-the-major-industries/}{fast growing industry itself}, and enhance more and more automation in classical industry as well.
-	
-	Among others, popular industries are:
-	\begin{itemize}
-		\item Embedded systems,
-		\item Mobility, and
-		\item Digital assistants
-	\end{itemize}
-	
-	Most applications are of the supervised type.
-	
-	The demand for highly skilled ML engineers is growing correspondingly.
-}
-
-\frame{\frametitle{Instances of ML applications}
-	\begin{itemize}
-		\item Recommendation systems
-		\begin{itemize}
-			\item Which ads to display on a website?
-			\item Which items are \href{https://tech.instacart.com/deep-learning-with-emojis-not-math-660ba1ad6cdc}{most likely put into cart next} by the user?
-		\end{itemize}\pause
-		\item Forecasting
-		\begin{itemize}
-			\item Weather, sales, \href{https://eng.uber.com/forecasting-introduction/}{geospatial Uber calls}, \href{https://www.kaggle.com/c/recruit-restaurant-visitor-forecasting}{restaurant}/\href{https://towardsdatascience.com/web-traffic-forecasting-f6152ca240cb}{website} traffic
-			\item Material attrition in engineering processes (predictive maintenance)
-		\end{itemize}\pause
-		\item Classification/Regression
-		\begin{itemize}
-			\item Speech assistants (Alexa/Siri), pedestrial detection (autonomous driving), \href{https://www.kaggle.com/c/severstal-steel-defect-detection}{fault detection in engineering processes}
-			\item \href{https://openai.com/blog/chatgpt}{large language models (LLM)}, credit scoring (fintech) 
-		\end{itemize}\pause
-		\item \href{https://stablediffusionweb.com/}{Generative models}
-
-	\end{itemize}
-}
-
-%  Kaggle comps
-\frame{\frametitle{ML competitions with price pool}
-	\begin{figure}
-		\includegraphics[width=9cm]{fig/lec08/comp_logos.pdf}
-		\caption{Kaggle and DrivenData}
-		\label{fig:comp_logos}
-	\end{figure}
-	
[…]
