### 2025-06-02 fa82de6
minor corrections to lec09
