+++ lecture/tex/Lecture09.tex
@@ -23,7 +23,7 @@ Until further notice we assume that:
 Focus of this and the next lecture:
 \begin{itemize}
 	\item Transferring previous RL methods from discrete to continuous state-space problems in the on-policy case.\pause
-	\item Applying off-policy approaches with function approximation is not straightforward and will be largely skipped.  
+	\item Applying off-policy prediction with function approximation is not straightforward and will be largely skipped.  
 	\begin{itemize}
 		\item For further insights we refer to chapter 11 in \textit{R. Sutton and G. Barto, Reinforcement learning: an introduction, 2018}. 
 	\end{itemize}
@@ -94,7 +94,7 @@ Due to the usage of function approximation one incremental learning step changes
 	\end{itemize}
 	\item Due to generalization impact we need to define an accuracy metric on the entire state space (the RL prediction goal):
 \end{itemize}
-\begin{defi}{Mean Squared Value Error}{VE}
+\begin{defi}{Mean squared value error}{VE}
 The RL prediction objective is defined as the mean squared value error
 \begin{equation}
 	\overline{\mbox{VE}}(\bm{w}) = \int_{\mathcal{X}}\mu(\bm{x})\left[v_{\pi}(\bm{x}) - \hat{v}(\tilde{\bm{x}}, \bm{w})\right]^2
@@ -103,7 +103,7 @@ The RL prediction objective is defined as the mean squared value error
 with $\mu(\bm{x})\in\left\{\mathbb{R}|\mu(\bm{x})\geq 0\right\}$ being a state distribution weight with $\int_{\mathcal{X}}\mu=1$. 
 \end{defi}\pause
 \begin{itemize}
-	\item Practical note: As the true value $v_{\pi}(\bm{x})$ is most likely unknown in most tasks, \eqref{eq:VE} cannot be computed exactly but only estimated.
+	\item Practical note: as the true value $v_{\pi}(\bm{x})$ is most likely unknown in most tasks, \eqref{eq:VE} cannot be computed exactly but only estimated.
 \end{itemize}
 }
 
@@ -170,11 +170,11 @@ Second challenge:
 %% Updating the Parameter Vector to Find (Local) Optimum %%
 %%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%
 \frame{\frametitle{Updating the parameter vector to find (local) optimum}
-Transferring the idea of incremental learning steps from the tabular case 
+Transferring the idea of \hl{incremental learning steps} from the tabular case 
 \begin{equation}
 	\hat{v}(x) \leftarrow \hat{v}(x) + \alpha \left[v_\pi(x) - \hat{v}(x)\right] 
 \end{equation}
-to function approximation using a gradient descent update:
+to function approximation using a \hl{gradient descent update}:
 \begin{equation}
 	\bm{w}\leftarrow\bm{w}-\alpha \nabla_{\bm{w}}J(\bm{w}).
 	\label{eq:gradient_learning}
@@ -203,7 +203,7 @@ to function approximation using a gradient descent update:
 \begin{column}{0.54\textwidth}
 \begin{minipage}[c]{\linewidth}
 \begin{itemize}
-	\item Full calculus of $\nabla_{\bm{w}}J(\bm{w})$:
+	\item Full batch calculus of $\nabla_{\bm{w}}J(\bm{w})$:
 	\begin{itemize}
[…]
