### 2025-06-25 1dcd6e3
typo corr to lec10/11
