+++ lecture/tex/Lecture11.tex
@@ -15,7 +15,7 @@ to (direct) policy-based solutions:
 \end{equation}\pause
 \begin{itemize}
 	\item Above, $\bm{\theta}\in \mathbb{R}^d$ is the policy parameter vector.
-	\item Note, that $\bm{u}$ is now vectorial and might contain multiple continuous quantities. 
+	\item Note, that $\bm{u}$ is vectorial and might contain multiple continuous quantities now. 
 \end{itemize}
 \pause
 \begin{block}{Goal of today's lecture}
@@ -45,7 +45,7 @@ to (direct) policy-based solutions:
 %% Motivating Example (2): Strategic Gaming %%
 %%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%
 \frame{\frametitle{Motivating example: strategic gaming}
-\onslide<1->Task: Two-player game of extended rock-paper-scissors 
+\onslide<1->Task: two-player game of extended rock-paper-scissors 
 \begin{itemize}
 	\onslide<2->\item A deterministic policy (i.e., value-based with given feature representation) can be easily exploited by the opponent.
 	\onslide<3->\item	Conversely, a uniform random policy would be unpredictable.
@@ -200,7 +200,7 @@ Possible optimality metrics:
 		\item multidimensional and
 		\item non-stationary problem.
 	\end{itemize}}
-	\onslide<3->{\item Hence, we might consider global optimization techniques\footnote[1]{\onslide<3->{Recommended reading: J. Stork et al., \textit{A new Taxonomy of Continuous Global Optimization Algorithms}, \href{https://arxiv.org/abs/1808.08818}{https://arxiv.org/abs/1808.08818}, 2020}} like
+	\onslide<3->{\item Hence, we might consider global optimization techniques\footnote[1]{\onslide<3->{Recommended reading: J. Stork et al., \textit{A new Taxonomy of Continuous Global Optimization Algorithms}, Natural Computing 21, pp. 219-242, \href{https://doi.org/10.1007/s11047-020-09820-4}{https://doi.org/10.1007/s11047-020-09820-4}, 2022}} like
 		\begin{itemize}
 		\item Simple heuristics: random search, grid search,...}
 		\onslide<4->\item Meta-heuristics: evolutionary algorithms, particle swarm,.... 
