### 2025-06-30 #226 @ethanwhite
Hi @XyDrKRulof - Apologies for the long delay since your initial submission. Please let me know if you're still interested in having this paper reviewed at JOSE. If you are I work on getting the review process started ASAP
