### 2025-07-05 c30ea65
update on new course repo template
