+++ lecture/tex/Lecture01.tex
@@ -11,24 +11,22 @@
 %%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%
 \begin{frame}
 	\frametitle{The teaching team}
-	\vspace{-0.5cm}
-	\begin{columns}[t]
-	
-		\column[T]{0.14\textwidth}
+	\centering
+	\begin{minipage}[t]{0.27\linewidth}
+		\centering
 		\begin{figure}
-			\centering
-				\includegraphics[height=2.75cm]{fig/lec01/Abdelwanis.jpg}
-				\caption*{Ali \\Abdelwanis}
+			\includegraphics[height=2.75cm]{fig/lec01/Abdelwanis.jpg}
+			\caption*{Ali \\Abdelwanis}
 		\end{figure}
-			
-		\column[T]{0.14\textwidth}
+	\end{minipage}
+	\hspace{1cm}
+	\begin{minipage}[t]{0.27\linewidth}
+		\centering
 		\begin{figure}
-			\centering
-				\includegraphics[height=2.75cm]{fig/lec01/Wallscheid.jpg}
-				\caption*{Oliver\\ Wallscheid}
+			\includegraphics[height=2.75cm]{fig/lec01/Wallscheid.jpg}
+			\caption*{Oliver\\ Wallscheid}
 		\end{figure}
-		
-	\end{columns}
+	\end{minipage}
 	\vspace{-0.25cm}
 	\begin{varblock}{Contact}
 		\begin{itemize}
