### 2025-07-07 7072322
small changes of animation order
