### 2025-07-28 c61c0cf
Refinement (#22)
* adding course template repo as submodule

* first step towards USI refactor

* course template update

* team update

* further refactoring to Siegen template

* course template update

* test build process

* furhter refactoring input

* update template

* changed the foldername whaddup

* further changes due to folder renaming

* Update LICENSE

* delete depreciated files

* changed the foldername

* readme refactoring

* action test

* adding automated section labels for pdf destinations

* addin automated pdf labels / destinations

* readme update

* further readme updat
[…]
