+++ exercises/templates/ex05/solution.py
@@ -1,207 +1,0 @@
-import numpy as np
-import random
-
-
-class RaceTrackEnv:
-    """
-    RaceTrackEnv object maintains and updates the race track 
-    state. Interaction with the class is through
-    the step() method (see OpenAI Gym interface)
-    
-    The class constructor is given a race course as a list of 
-    strings. The constructor loads the course and initializes 
-    the environment state.
-    """
-    
-    lbl2int = {'o': 1, '-': 0, '+': 2, 'W': -1}
-    
-    def __init__(self, course):
-        """
-        Load race course, set any min or max limits in the 
-        environment (e.g. max speed), and set initial state.
-        Initial state is random position on start line with 
-        velocity = (0, 0).
-        
-        Example:
-            tiny_right_turn_course = 
-                  ['WWWWWW',
-                   'Woooo+',
-                   'Woooo+',
-                   'WooWWW',
-                   'WooWWW',
-                   'WooWWW',
-                   'WooWWW',
-                   'W--WWW',]
-        
-        Args:
-            course: List of text strings used to construct
-                race-track.
-                    '-': start line
-                    '+': finish line
-                    'o': track
-                    'W': wall
-        
-        Returns:
-            self
-        """
-        self.MAX_VELOCITY = 5
-        self.start_positions = []
-        self.course = None
-        self._load_course(course)
-        self._random_start_position()
-        self.velocity = np.array([0, 0], dtype=np.int8)
-        self.bounds = (len(course), len(course[0]))
-
-
-    def step(self, action):
-        """
-        Perform given action on the environment.
-        The reward for every action and state is -1, except for when 
-        reaching the finish line.
-        
-        Args:
-            action: integer-encoded {0-8} or a 
-                2-tuple (a_y, a_x) in {[-1, 0, 1], [-1, 0, 1]}.
-        
-        Returns:
-            5-tuple: (new_state, reward, terminated, truncated, info)
-                new_state: the new state resulting from taken action
-                reward: The reward obtained from taken action
-                terminated: Bool. Whether we are in a terminal state.
-                truncated: Bool. Whether we have to reset, but not finished the race track.
-                Info: Dict. Arbitrary information.
-                
-        """
-        ### BEGIN SOLUTION
-        if not isinstance(action, tuple):
-            action = self.action_to_tuple(action)
-        
-        reward = -1.0
-        terminated = False
-        truncated = False
-        if self.is_terminal_state():
-            reward = 0
-            terminated = True
-        else:
-            # update velocity
-            self.velocity += np.array(action, dtype=np.int8)
-            self.velocity = np.clip(self.velocity, -self.MAX_VELOCITY, self.MAX_VELOCITY)
-            # project jump to new position virtually
-            projected_path_y = np.clip(
-                np.around(np.linspace(self.position[0], 
-                                      self.position[0] + self.velocity[0],
-                                      self.MAX_VELOCITY*2),
-                          decimals=1).astype(int), 
-                0, self.bounds[0]-1)
-            projected_path_x = np.clip(
-                np.around(np.linspace(self.position[1], 
-                                      self.position[1] + self.velocity[1],
-                                      self.MAX_VELOCITY*2),
-                          decimals=1).astype(int),
-                0, self.bounds[1]-1)
-            projected_steps = list(self.course[projected_path_y, projected_path_x])
-            # hits finish line?
-            if self.lbl2int['+'] in projected_steps:                
-                # has grass been hit before?
-                if self.lbl2int['W'] in projected_steps:
-                    s = self.reset()
[…]
