+++ paper/paper.bib
@@ -1,0 +1,279 @@
+@article{Sutton2005ReinforcementLA,
+  title={Reinforcement Learning: An Introduction},
+  author={Richard S. Sutton and Andrew G. Barto},
+  journal={IEEE Transactions on Neural Networks},
+  year={2005},
+  volume={16},
+  pages={285-286},
+  url={https://api.semanticscholar.org/CorpusID:9166388}
+}
+
+@Misc{silver2015,
+  author = {David Silver},
+  title = {Lectures on Reinforcement Learning},
+  howpublished = {\textsc{url:}~\url{https://www.davidsilver.uk/teaching/}},
+  year = {2015}
+}
+
+@conference{Kluyver2016jupyter,
+  Title = {Jupyter Notebooks -- a publishing format for reproducible computational workflows},
+  Author = {Kluyver, Thomas and Ragan-Kelley, Benjamin and P{\'e}rez, Fernando et al},
+  Booktitle = {Positioning and Power in Academic Publishing: Players, Agents and Agendas},
+  Editor = {F. Loizides and B. Schmidt},
+  Organization = {IOS Press},
+  Pages = {87 - 90},
+  Year = {2016}
+}
+
+@software{reback2020pandas,
+  author       = {pandas},
+  title        = {pandas-dev/pandas: Pandas},
+  month        = feb,
+  year         = 2020,
+  publisher    = {Zenodo},
+  version      = {latest},
+  doi          = {10.5281/zenodo.3509134},
+  url          = {https://doi.org/10.5281/zenodo.3509134}
+}
+
+@InProceedings{ mckinney-proc-scipy-2010,
+  author    = { {W}es {M}c{K}inney },
+  title     = { {D}ata {S}tructures for {S}tatistical {C}omputing in {P}ython },
+  booktitle = { {P}roceedings of the 9th {P}ython in {S}cience {C}onference },
+  pages     = { 56 - 61 },
+  year      = { 2010 },
+  editor    = { {S}t\'efan van der {W}alt and {J}arrod {M}illman },
+  doi       = { 10.25080/Majora-92bf1922-00a }
+}
+
+@misc{towers_gymnasium_2023,
+  title = {Gymnasium},
+  url = {https://zenodo.org/record/8127025},
+  abstract = {An API standard for single-agent reinforcement learning environments, with popular reference environments and related utilities (formerly Gym)},
+  urldate = {2023-07-08},
+  publisher = {Zenodo},
+  author = {Towers, Mark and Terry, Jordan K. and Kwiatkowski, Ariel et al},
+  month = mar,
+  year = {2023},
+  doi = {10.5281/zenodo.8127026},
+}
+
+@misc{tensorflow2015-whitepaper,
+  title={ {TensorFlow}: Large-Scale Machine Learning on Heterogeneous Systems},
+  url={https://www.tensorflow.org/},
+  note={Software available from tensorflow.org},
+  author={
+    Abadi, Mart\'{i}n and
+    Agarwal,Ashish and
+    Barham, Paul et al},
+  year={2015},
+}
+
+@incollection{NEURIPS2019_9015,
+  title = {PyTorch: An Imperative Style, High-Performance Deep Learning Library},
+  author = {Paszke, Adam and Gross, Sam and Massa, Francisco et al},
+  booktitle = {Advances in Neural Information Processing Systems 32},
+  pages = {8024--8035},
+  year = {2019},
+  publisher = {Curran Associates, Inc.},
+  url = {http://papers.neurips.cc/paper/9015-pytorch-an-imperative-style-high-performance-deep-learning-library.pdf}
+} 
+
+@misc{gym,
[…]
