### 2025-08-01 #226 @XyDrKRulof
@ethanwhite we are still interested in having our paper reviewed at JOSE. I just updated the paper to match its contents with the work done on the repository since the initial submission.
