### 2025-08-21 6d9b221
Added Paderborn theme back
