+++ lecture/main.tex
@@ -1,4 +1,4 @@
-\documentclass[CC-BY-NC-SA]{../course_template/lectureClass}
+\documentclass[CC-BY-NC-SA, unipaderborn]{../course_template/lectureClass}
 \makeglossaries % dummy generate to avoid glossary package error
 \title{Reinforcement learning}
 \author{Oliver Wallscheid}
