### 2025-08-21 40a2f0b
Update README.md
Changed lecture links pointing to Siegen repository back to Paderborn repository
