### 2025-08-21 fbb6e00
Update README.md
Fixed error in lecture links
