### 2025-08-21 #226 @ethanwhite
:wave: @freddiev4 @gcdeshpande @kbodwin - would you be willing to review this submission for JOSE? We carry out our checklist-driven reviews here in GitHub issues and follow these guidelines: https://openjournals.readthedocs.io/en/jose/review_criteria.html
