### 2025-09-02 #306 @editorialbot
[REVIEW]: Reinforcement Learning: A Comprehensive Open-Source Course
**Submitting author:** @XyDrKRulof (<a href="http://orcid.org/0009-0002-1576-2465">Darius Jakobeit</a>)
**Repository:** https://github.com/upb-lea/reinforcement_learning_course_materials
**Branch with paper.md** (empty if default branch):
**Version:** 2025_v2
