### 2025-09-08 upb-lea/reinforcement_learning_course_materials#24 @arokem
[JOSE paper review] Instructions for building the slides
There are no instructions for how to build the latex slides from the source.

**Describe the solution you'd like**

A README file here: https://github.com/upb-lea/reinforcement_learning_course_materials/tree/master/lecture, with some instructions on how to operate the source of the slides would be helpful.
