### 2025-09-08 upb-lea/reinforcement_learning_course_materials#25 @arokem
[JOSE paper review] Contributions section is bare-bones
I noticed [this section](https://github.com/upb-lea/reinforcement_learning_course_materials/tree/master?tab=readme-ov-file#contributions) in the README. While it meets the minimal standard, it would be nice to flesh this out a bit more. In particular, how does one contact the authors (in the statement "please contact us")? Ideally, contacting the authors would be done through a public channel (e.g., through GitHub issues), rather than private email, so that there is transparency about the development process. But this not a strict requirement, so I leave it up to the authors.
