### 2025-09-08 #306 @arokem
I have completed my review, checking all but one of the boxes. I had only a couple of small issues/suggestions:

https://github.com/upb-lea/reinforcement_learning_course_materials/issues/25
https://github.com/upb-lea/reinforcement_learning_course_materials/issues/24

In addition, I have a small language/usage suggestion: I believe that the "Corona pandemic" (L32 of the rendered pdf) is better known in English as the "COVID-19 pandemic".
