### 2025-09-10 59a7272
Revise contributions section in README (issue #25)
Extended contributions section in README to emphasize that contact with us should be made through GitHub issues.
