### 2025-09-10 a0a6f11
Added a README.md for the lecture build process. Additionally fixed the tidy up process of the build.py to remove binaries after the build process and to move the built pdf into a subfolder instead of the main folder (issue #24)
