### 2025-09-10 1f43922
Changed corona pandemic to COVID-19 pandemic
