+++ paper/paper.md
@@ -67,7 +67,7 @@ The course spans 14 weeks, comprising 14 lectures and 12 exercises. Accompanying
 
 The lecture follows Richard S. Sutton and Andrew G. Barto's fundamentals book on reinforcement learning [@Sutton2005ReinforcementLA] and takes inspiration from the reinforcement learning lecture script delivered by David Silver [@silver2015]. The exercises are programmed in Python using Jupyter notebooks [@Kluyver2016jupyter] for presentation. Important libraries for machine and reinforcement learning are introduced, such as pandas [@reback2020pandas][@mckinney-proc-scipy-2010], gymnasium [@towers_gymnasium_2023], PyTorch [@NEURIPS2019_9015], scikit-learn [@sklearn_api], and stable-baselines3 [@stable-baselines3]. 
 
-The authors of this course have experience working with reinforcement learning in the domain of electrical engineering, in particular in electric drive [@deep_q_torque] and grid control [@10182718]. The course has first been held under the constraints of the Corona pandemic in 2020, resorting to an online, asynchronous learning experience. It has been extended with a session on more contemporary algorithms in 2022. In subsequent years the course has been revised to incorporate experience from teaching the course and to align the structure of the exercises. All versions (for each year's revision) are available inside the publicly available [GitHub repository](https://github.com/upb-lea/reinforcement_learning_course_materials).
+The authors of this course have experience working with reinforcement learning in the domain of electrical engineering, in particular in electric drive [@deep_q_torque] and grid control [@10182718]. The course has first been held under the constraints of the COVID-19 pandemic in 2020, resorting to an online, asynchronous learning experience. It has been extended with a session on more contemporary algorithms in 2022. In subsequent years the course has been revised to incorporate experience from teaching the course and to align the structure of the exercises. All versions (for each year's revision) are available inside the publicly available [GitHub repository](https://github.com/upb-lea/reinforcement_learning_course_materials).
 
 # Statement of Need
 
