### 2025-09-10 upb-lea/reinforcement_learning_course_materials#25 @arokem
Looks good to me! Small suggestion, which you can take or leave is to use this link: https://github.com/upb-lea/reinforcement_learning_course_materials/issues/new/choose for creating new issues, rather than the `../../issues` that you are currently using. But that's a small thing, so I'll go ahead and close this issue.
