### 2025-09-10 #306 @arokem
OK - I just checked the last box based on resolution of these issues. I am happy with the paper/repo as is. Let me know if I need to do anything else to indicate my 👍
