### 2025-09-19 #306 @sconde
I’d like to wrap up my review, but there’s still an open issue that hasn’t been addressed. Could we resolve that before I finalize my feedback?
