### 2025-09-19 #306 @ethanwhite
Thanks @sconde!

@XyDrKRulof - it looks like we're ready to wrap up the review as soon as you address https://github.com/upb-lea/reinforcement_learning_course_materials/issues/26. Let me now if you have any questions as you proceed.
