### 2025-09-19 #306 @XyDrKRulof
Thanks @sconde and @arokem for your reviews :)
I am currently waiting for the answer of the colleague that has last updated the exercises. Then I can set the versions in the requirements.txt of the repository to their latest states. The issue should be resolved by Monday, sorry for the delay on my side.
