### 2025-09-22 f784f50
Added versions for libraries that had no specified version (see #26)
