+++ requirements.txt
@@ -2,13 +2,13 @@ numpy==2.1.0
 matplotlib==3.10.1
 scipy==1.15.2
 pandas==2.2.3
-openpyxl
-seaborn==0.13.2 
+openpyxl==3.1.5
+seaborn==0.13.2
 tqdm==4.67.1
 jupyter==1.1.1
 gymnasium==1.1.1
 pygame==2.6.1
 scikit-learn==1.6.1
-torch
+torch==2.6.0
 stable-baselines3 == 2.6.0
-Box2D 
+Box2D==2.3.10
