### 2025-09-22 #306 @sconde
The repository is well organized, and the course is clear, thorough, and educational. It covers reinforcement learning topics with a good balance of theory and practice, and the documentation supports smooth progression.
