### 2025-09-26 #306 @ethanwhite
Thanks so much to @arokem and @sconde for the excellent reviews and to @XyDrKRulof to addressing the recommendations so quickly! This paper is a great reminder of what positive, constructive, collaborative, peer review can look like!

I'll be going through editor checks and will be in touch with @XyDrKRulof about any last things that need to be done before we formally accept the paper.
