### 2025-09-26 #306 @ethanwhite
@XyDrKRulof here are the last couple of things we need to finish up:
1. Remove the https://doi.org/ from the DOI in the CORONATO2020101964 entry in the paper.bib file
2. In paper.md, when there are multiple citations in a row, e.g. `[@atari-first] [@starcraft2]`, use one set of `[]` and separate the citations with `;`, e.g., `[@atari-first; @starcraft2]`. This will result in properly formatted inline citations.
3. The alternative courses cited in the second paragraph of the Statement of Need should be cited as websites with a reference in the .bib file like you did for Silver 2015.
4. Consider making a few editorial changes
  - 1st sentence of Statement of Need: "in both, academia as well as popular science media." -> "in both academia and popular science media" or "in academia as well as popular science media"
  - Second paragraph of Content: "which are lend from third party open-source libraries" I'm not sure what is intended by "lend" in this context. Possibly "learned" or "taken" or "implemented in"?
  - Conclusion: "Both, the lecture content and the exercises, are" -> " Both the lecture content and the exercises are"
5. Once you have completed the above changes please make a new release and archive it to Zenodo. I find the easiest way to do this is through the [GitHub - Zenodo integration](https://docs.github.com/en/repositories/archiving-a-github-repository/referencing-and-citing-content) in which case you will need to setup the integration first and then make the new release in GitHub. Alternatively it is fine to manually upload a copy of the repository to Zenodo. The author line and title on Zenodo should match the paper and the release tag should match the one on GitHub. One way to do this automatically with the integration is to add a [`.zenodo.json` file](https://help.zenodo.org/docs/github/describe-software/zenodo-json/) to the GitHub repository prior to the release. If you have any questions about this step let me know.
6. Once all the changes are made and the new release and Zenodo archive are created please respond to this issue with the DOI for the Zenodo archive and the tag for the GitHub release.
