### 2025-10-13 33303ef
Editorial changes for the JOSE paper as requested by Ethan White
