+++ paper/paper.md
@@ -65,15 +65,15 @@ We present an open-source repository of an extensive course on reinforcement lea
 
 The course spans 14 weeks, comprising 14 lectures and 12 exercises. Accompanying video materials from real lectures and exercises are provided to aid in understanding the course content. They are available on a [YouTube channel](https://www.youtube.com/@uni_pb_lea/playlists) under an Creative Commons license. The open-source nature of the course allows other teachers to freely adapt the materials for their own teaching purposes. The primary goal is to equip learners with a solid theory of reinforcement learning principles, as well as the practical tools to solve real-world engineering problems from different domains, such as electrical engineering.
 
-The lecture follows Richard S. Sutton and Andrew G. Barto's fundamentals book on reinforcement learning [@Sutton2005ReinforcementLA] and takes inspiration from the reinforcement learning lecture script delivered by David Silver [@silver2015]. The exercises are programmed in Python using Jupyter notebooks [@Kluyver2016jupyter] for presentation. Important libraries for machine and reinforcement learning are introduced, such as pandas [@reback2020pandas][@mckinney-proc-scipy-2010], gymnasium [@towers_gymnasium_2023], PyTorch [@NEURIPS2019_9015], scikit-learn [@sklearn_api], and stable-baselines3 [@stable-baselines3]. 
+The lecture follows Richard S. Sutton and Andrew G. Barto's fundamentals book on reinforcement learning [@Sutton2005ReinforcementLA] and takes inspiration from the reinforcement learning lecture script delivered by David Silver [@silver2015]. The exercises are programmed in Python using Jupyter notebooks [@Kluyver2016jupyter] for presentation. Important libraries for machine and reinforcement learning are introduced, such as pandas [@reback2020pandas; @mckinney-proc-scipy-2010], gymnasium [@towers_gymnasium_2023], PyTorch [@NEURIPS2019_9015], scikit-learn [@sklearn_api], and stable-baselines3 [@stable-baselines3]. 
 
 The authors of this course have experience working with reinforcement learning in the domain of electrical engineering, in particular in electric drive [@deep_q_torque] and grid control [@10182718]. The course has first been held under the constraints of the COVID-19 pandemic in 2020, resorting to an online, asynchronous learning experience. It has been extended with a session on more contemporary algorithms in 2022. In subsequent years the course has been revised to incorporate experience from teaching the course and to align the structure of the exercises. All versions (for each year's revision) are available inside the publicly available [GitHub repository](https://github.com/upb-lea/reinforcement_learning_course_materials).
 
 # Statement of Need
 
-Recent developments in (deep) reinforcement learning caused considerable excitement in both, academia as well as [popular science media](https://www.youtube.com/watch?v=WXuK6gekU1Y). Starting with beating champions in complex board games such as chess [@chess] and Go [@SilverHuangEtAl16nature], breaking human records in a wide variety of video games [@atari-first] [@starcraft2], up to recent solutions in real-world (control) applications [@robotics] [@rlhf] [@motors] [@zejnullahu2022applications] [@CORONATO2020101964] [@traffic], reinforcement learning agents have been proven to be a control or decision-making solution for a wide variety of application domains. Reinforcement learning poses an elegant and data-driven path to a control solution with minimal expert knowledge involved, which makes it highly attractive for many different research domains. A similar development has already been observed in recent years with regard to deep supervised learning. 
[…]
