+++ paper/paper.bib
@@ -12,7 +12,31 @@
   author = {David Silver},
   title = {Lectures on Reinforcement Learning},
   howpublished = {\textsc{url:}~\url{https://www.davidsilver.uk/teaching/}},
-  year = {2015}
+}
+
+@misc{deep-rl-course,
+  author = {Simonini, Thomas and Sanseviero, Omar},
+  title = {The Hugging Face Deep Reinforcement Learning Class},
+  year = {2023},
+  publisher = {GitHub},
+  journal = {GitHub repository},
+  howpublished = {\url{https://github.com/huggingface/deep-rl-class}},
+}
+
+@Misc{stanford,
+  author = {Emma Brunskill},
+  title = { CS234: Reinforcement Learning Winter 2025 },
+  howpublished = {\textsc{url:}~\url{https://web.stanford.edu/class/cs234/}},
+  year = {2025},
+
+}
+
+@Misc{spinningup,
+  author = {Josh Achiam},
+  title = { Spinning Up in Deep Reinforcement Learning },
+  howpublished = {\textsc{url:}~\url{https://spinningup.openai.com/}},
+  year = {2018},
+
 }
 
 @conference{Kluyver2016jupyter,
@@ -223,7 +247,7 @@
   pages = {101964},
   year = {2020},
   issn = {0933-3657},
-  doi = {https://doi.org/10.1016/j.artmed.2020.101964},
+  doi = {10.1016/j.artmed.2020.101964},
   url = {https://www.sciencedirect.com/science/article/pii/S093336572031229X},
   author = {Antonio Coronato and Muddasar Naeem and Giuseppe {De Pietro} and Giovanni Paragliola},
   keywords = {Artificial intelligence, Reinforcement learning, Healthcare, Personalized medicine},
