+++ paper/paper.md
@@ -71,9 +71,9 @@ The authors of this course have experience working with reinforcement learning i
 
 # Statement of Need
 
-Recent developments in (deep) reinforcement learning caused considerable excitement in both academia and [popular science media](https://www.youtube.com/watch?v=WXuK6gekU1Y). Starting with beating champions in complex board games such as chess [@chess] and Go [@SilverHuangEtAl16nature], breaking human records in a wide variety of video games [@atari-first; @starcraft2], up to recent solutions in real-world (control) applications [@robotics; @rlhf; @motors] [@zejnullahu2022applications; @CORONATO2020101964; @traffic], reinforcement learning agents have been proven to be a control or decision-making solution for a wide variety of application domains. Reinforcement learning poses an elegant and data-driven path to a control solution with minimal expert knowledge involved, which makes it highly attractive for many different research domains. A similar development has already been observed in recent years with regard to deep supervised learning. 
+Recent developments in (deep) reinforcement learning caused considerable excitement in both academia and [popular science media](https://www.youtube.com/watch?v=WXuK6gekU1Y). Starting with beating champions in complex board games such as chess [@chess] and Go [@SilverHuangEtAl16nature], breaking human records in a wide variety of video games [@atari-first; @starcraft2], up to recent solutions in real-world (control) applications [@robotics; @rlhf; @motors; @zejnullahu2022applications; @CORONATO2020101964; @traffic], reinforcement learning agents have been proven to be a control or decision-making solution for a wide variety of application domains. Reinforcement learning poses an elegant and data-driven path to a control solution with minimal expert knowledge involved, which makes it highly attractive for many different research domains. A similar development has already been observed in recent years with regard to deep supervised learning. 
 
-An increasing amount of educational resources is available due to the traction RL has gained in recent years. However, most courses lack either the continuity of topics ranging from the foundations up to the advanced topics of deep reinforcement learning, practical programming exercises accompanying each theoretical lecture, the testing at university level, or free availability. Alternative courses often focus on games [deep-rl-course] or a mix of theoretical and practical questions for their exercises [@stanford; spinningup]. In contrast, our course utilizes practical application scenarios from a wide variety of domains with a strong focus on classical control engineering tasks. This course can therefore help accelerate establishing reinforcement learning solutions within real-world applications.
+An increasing amount of educational resources is available due to the traction RL has gained in recent years. However, most courses lack either the continuity of topics ranging from the foundations up to the advanced topics of deep reinforcement learning, practical programming exercises accompanying each theoretical lecture, the testing at university level, or free availability. Alternative courses often focus on games [@deep-rl-course] or a mix of theoretical and practical questions for their exercises [@stanford; @spinningup]. In contrast, our course utilizes practical application scenarios from a wide variety of domains with a strong focus on classical control engineering tasks. This course can therefore help accelerate establishing reinforcement learning solutions within real-world applications.
 
 
 # Target Audience and Learning Goals
@@ -129,7 +129,7 @@ Table: Summary of course exercises.
 | 11 | Moon Landing with Actor-Critic Methods |
 | 12 | Shoot for the moon with DDPG & PPO |
 
-Lectures and exercises which share the same number also deal with the same topics. Thus, theoretical basics are provided in the lecture, which are to be implemented and evaluated in the exercises on the basis of specific application examples which are taken from third party open-source libraries [@gym][@towers_gymnasium_2023]. This allows the learners to internalize learned contents practically. However, the lecture can be studied independently of the exercises and the exercises independently of the lecture in case of self-learning.
[…]
