### 2025-10-13 f76ecdb
Re-added David Silver release year
