+++ paper/paper.bib
@@ -11,6 +11,7 @@
 @Misc{silver2015,
   author = {David Silver},
   title = {Lectures on Reinforcement Learning},
+  year = {2015},
   howpublished = {\textsc{url:}~\url{https://www.davidsilver.uk/teaching/}},
 }
 
