+++ .github/workflows/buildPDFs.yml
@@ -29,7 +29,7 @@ jobs:
         uses: actions/upload-pages-artifact@v3 #actions/upload-artifact@v4
         with:
           name: github-pages
-          path: built/
+          path: lecture/built/
 
 
   deploy:
