### 2025-10-14 #306 @XyDrKRulof
Hey @ethanwhite,

I have successfully worked through your to-do list.

DOI: 10.5281/zenodo.17347442
GitHub tag: 2025_v1

Let me know if you need anything else from my side.
