### 2025-10-29 #306 @ethanwhite
Thanks for taking care of this @XyDrKRulof! Everything looks great! I'm going to officially recommend acceptance, which will move this to the Editor In Chief's desk for final approval and publication. It's been great working with you on this paper and the OER is a real contribution to the field!
