### 2025-10-29 #306 @editorialbot
:wave: @openjournals/jose-eics, this paper is ready to be accepted and published.
