### 2025-12-09 #306 @labarba
The repo shows a CC-BY-NC-SA license, but the Zenodo deposit shows CC-BY.

While you are looking into this, have you considered carefully the non-commercial clause? Some folks choose this from gut feeling, but may not realize the complications it brings, that's why I ask. (Have a look at my licensing tutorial, where I [address this](https://gw-ospo.github.io/oss-licensing/license_your_own_work.html#what-about-a-non-commercial-cc-license).)
