### 2026-01-05 ef64ef1
Update README.md
Changed License in README to CC BY 4.0
