+++ README.md
@@ -1,7 +1,7 @@
 # Reinforcement learning course
 
 [![Build Status](https://github.com/upb-lea/reinforcement_learning_course_materials/actions/workflows/buildPDFs.yml/badge.svg)](https://github.com/upb-lea/reinforcement_learning_course_materials/actions/workflows/buildPDFs.yml)
-[![CC BY 4.0][cc-by-shield]][cc-sa]
+[![CC BY 4.0][cc-by-shield]][cc-by]
 [![made-with-python](https://img.shields.io/badge/Made%20with-Python-1f425f.svg)](https://www.python.org/)
 [![made-with-latex](https://img.shields.io/badge/Made%20with-LaTeX-1f425f.svg)](https://www.latex-project.org/)
 
