### 2026-01-05 #306 @XyDrKRulof
Hey @labarba, happy new year and sorry for the late reply. We discussed the licensing internally and decided to change the license of the repo to CC-BY. I was unsure whether I should create a new Zenodo release but as it contained the old README which stated the CC-BY-NC-SA license I did a new one:

DOI: 10.5281/zenodo.18153154

The release version is now _2025_v2_.

If this was unnecessary then we could keep the older release (2025_v1). Apart from the licensing, nothing changed. Please let me know if there is something else I should do now.
