### 2026-03-28 #306 @labarba
I'm ready to go ahead and publish your paper. When you have a chance, do propagate the updated license information to slides and other places that it is mentioned.
