### 2026-03-28 #306 @labarba
Congratulations, @XyDrKRulof – your JOSE paper is published! 🚀

Big thanks to our Editor: @ethanwhite and Reviewers: @sconde, @arokem – you make this possible 🙏
